backend/requirements.txt
Flask
Flask-SQLAlchemy==3.0.2
flask-security-too
email_validator
bcrypt
flask_restful
flask_cors
flask-jwt-extended